(PC-2113): Add get_offerer_with_stock_count() to offerer repository
