+++ scripts/dashboard/diversification_statistics.py
@@ -4,25 +4,15 @@ from models import Offerer, UserOfferer, Venue, Offer, Stock, Booking, EventType
 from models.db import db
 from repository.booking_queries import count_all_used_booking
 from repository.offer_queries import get_active_offers_Ids_query
-from repository.offerer_queries import count_offerer
+from repository.offerer_queries import count_offerer, count_offerer_with_stock
 
 
 def get_offerer_count() -> int:
     return count_offerer()
 
 
-def _query_offerers_with_user_offerer():
-    return Offerer.query \
-        .join(UserOfferer) \
-        .distinct(Offerer.id)
-
-
 def get_offerer_with_stock_count() -> int:
-    return _query_offerers_with_user_offerer() \
-        .join(Venue, Venue.managingOffererId == Offerer.id) \
-        .join(Offer) \
-        .join(Stock) \
-        .count()
+    return count_offerer_with_stock()
 
 
 def get_offerers_with_offer_available_on_discovery_count() -> int:
