+++ tests/scripts/dashboard/diversification_statistics_test.py
@@ -15,100 +15,6 @@ from tests.test_utils import create_user, create_offerer, create_user_offerer, c
     create_offer_with_thing_product, create_venue, create_mediation, create_offer_with_event_product, create_booking
 
 
-class GetOffererWithStockCountTest:
-    def test_return_zero_if_no_offerer(self, app):
-        # When
-        number_of_offerers = get_offerer_with_stock_count()
-
-        # Then
-        assert number_of_offerers == 0
-
-    @clean_database
-    def test_return_1_if_offerer_with_user_offerer_and_stock(self, app):
-        # Given
-        user = create_user()
-        offerer = create_offerer()
-        user_offerer = create_user_offerer(user, offerer)
-        venue = create_venue(offerer)
-        offer = create_offer_with_thing_product(venue)
-        stock = create_stock(offer=offer)
-        PcObject.save(user_offerer, stock)
-
-        # When
-        number_of_offerers = get_offerer_with_stock_count()
-
-        # Then
-        assert number_of_offerers == 1
-
-    @clean_database
-    def test_return_0_if_offerer_without_user_offerer(self, app):
-        # Given
-        offerer = create_offerer()
-        venue = create_venue(offerer)
-        offer = create_offer_with_thing_product(venue)
-        stock = create_stock(offer=offer)
-        PcObject.save(offerer, stock)
-
-        # When
-        number_of_offerers = get_offerer_with_stock_count()
-
-        # Then
-        assert number_of_offerers == 0
-
-    @clean_database
-    def test_return_0_if_offerer_without_stock(self, app):
-        # Given
-        offerer = create_offerer()
-        user = create_user()
-        user_offerer = create_user_offerer(user, offerer)
-
-        PcObject.save(user_offerer)
-
-        # When
-        number_of_offerers = get_offerer_with_stock_count()
-
-        # Then
-        assert number_of_offerers == 0
-
-    @clean_database
-    def test_return_1_if_offerer_with_user_offerer_and_two_stock(self, app):
-        # Given
-        user = create_user()
-        offerer = create_offerer()
-        user_offerer = create_user_offerer(user, offerer)
-        venue = create_venue(offerer)
-        offer = create_offer_with_thing_product(venue)
-        stock1 = create_stock(offer=offer)
-        stock2 = create_stock(offer=offer)
-        PcObject.save(user_offerer, stock1, stock2)
-
-        # When
-        number_of_offerers = get_offerer_with_stock_count()
-
-        # Then
-        assert number_of_offerers == 1
-
-    @clean_database
-    def test_return_1_if_offerer_with_2_user_offerers_and_stock(self, app):
-        # Given
-        user1 = create_user()
-        user2 = create_user(email='other@email.com')
-        offerer = create_offerer()
-        user_offerer1 = create_user_offerer(user1, offerer)
-        user_offerer2 = create_user_offerer(user2, offerer)
-        venue = create_venue(offerer)
-        offer = create_offer_with_thing_product(venue)
-        stock1 = create_stock(offer=offer)
-        stock2 = create_stock(offer=offer)
-        PcObject.save(user_offerer1, user_offerer2, stock1, stock2)
-
-        # When
-        number_of_offerers = get_offerer_with_stock_count()
-
-        # Then
-        assert number_of_offerers == 1
-
-
 class GetOfferersWithOfferAvailableOnDiscoveryCountTest:
     @clean_database
     def test_returns_0_if_only_offerer_with_inactive_offer(self, app):
